\name{ahull.points}
\alias{ahull.points}
\title{ahull points}
\description{Calculate the points on the ashape.}
\usage{ahull.points(d, ..., 
    ahull = default.ahull(d))}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
  \item{ahull}{ahull }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





