\name{bottom.points}
\alias{bottom.points}
\title{bottom points}
\description{Positioning Method for the bottom of a group of points.}
\usage{bottom.points(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





