\name{bumpup}
\alias{bumpup}
\title{bumpup}
\description{Sequentially bump labels up, starting from the bottom, if they
collide with the label underneath.}
\usage{bumpup(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





