\name{calc.borders}
\alias{calc.borders}
\title{calc borders}
\description{Calculate bounding box based on newly calculated width and height.}
\usage{calc.borders(d, ...)}
\arguments{
  \item{d}{Data frame of point labels, with new widths and heights in the w
and h columns.}
  \item{\dots}{ignored.}
}



\author{Toby Dylan Hocking}





