\name{calc.boxes}
\alias{calc.boxes}
\title{calc boxes}
\description{Calculate boxes around labels, for collision detection.}
\usage{calc.boxes(d, debug = FALSE, 
    ...)}
\arguments{
  \item{d}{d }
  \item{debug}{debug }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





