\name{check.for.columns}
\alias{check.for.columns}
\title{check for columns}
\description{Stop if a data.frame does not have some columns.}
\usage{check.for.columns(d, 
    must.have)}
\arguments{
  \item{d}{data.frame to check.}
  \item{must.have}{column names to check.}
}



\author{Toby Dylan Hocking}





