\name{chull.points}
\alias{chull.points}
\title{chull points}
\description{Calculate the points on the convex hull.}
\usage{chull.points(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





