\name{closest.on.ahull}
\alias{closest.on.ahull}
\title{closest on ahull}
\description{Calculate closest point on the alpha hull with size of the boxes,
and put it outside that point.}
\usage{
closest.on.ahull(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





