\name{closest.on.chull}
\alias{closest.on.chull}
\title{closest on chull}
\description{Calculate closest point on the convex hull and put it outside that
point. Assume d is the center for each point cloud and then use
orig.data to calculate hull.}
\usage{
closest.on.chull(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





