\name{collide.up}
\alias{collide.up}
\title{collide up}
\description{Sequentially bump labels up, starting from the bottom, if they
collide with the label underneath.}
\usage{
collide.up(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}
