\name{compare.methods}
\Rdversion{1.1}
\alias{compare.methods}
\title{compare methods}
\description{Plot several label placement methods on the same page.}
\usage{
compare.methods(m, ..., horiz = FALSE)
}
\arguments{
  \item{m}{Vector of label placement function names.}
  \item{\dots}{Args to pass to dl}
  \item{horiz}{Arrange plots horizontally or vertically?}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}