\name{default.ahull}
\alias{default.ahull}
\title{default ahull}
\description{Calculate the default alpha parameter for ashape based on the
average size of label boxes.}
\usage{default.ahull(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





