\name{default.picker}
\alias{default.picker}
\title{default picker}
\description{Look at options() for a user-defined default Positioning Method
picker, and use that (or the hard-coded default picker), with the
calling environment to figure out a good default.}
\usage{default.picker(f)}
\arguments{
  \item{f}{Object class to look for (trellis or ggplot).}
}



\author{Toby Dylan Hocking}





