\name{defaultpf.ggplot}
\alias{defaultpf.ggplot}
\title{defaultpf ggplot}
\description{Default method selection method for ggplot2 plots.}
\usage{defaultpf.ggplot(geom, 
    p, L, colvar, ...)}
\arguments{
  \item{geom}{geom }
  \item{p}{p }
  \item{L}{L }
  \item{colvar}{colvar }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking}





