\name{defaultpf.lattice}
\alias{defaultpf.lattice}
\title{defaultpf lattice}
\description{If no Positioning Function specified, choose a default using this
function.}
\usage{
defaultpf.lattice()
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}
