\name{defaultpf.trellis}
\alias{defaultpf.trellis}
\title{defaultpf trellis}
\description{If no Positioning Method specified, choose a default using this
function. The idea is that this is called with all the variables
in the environment of \code{\link{panel.superpose.dl}}, and this can be
user-customizable by setting the directlabels.defaultpf.lattice
option to a function like this.}
\usage{defaultpf.trellis(lattice.fun.name, 
    groups, type, ...)}
\arguments{
  \item{lattice.fun.name}{lattice.fun.name }
  \item{groups}{groups }
  \item{type}{type }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





