\name{dens.gradient}
\alias{dens.gradient}
\title{dens gradient}
\description{Calculate a 2d density estimate then follow the gradient to a
point outside the convex hull.}
\usage{dens.gradient(d, ...)}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





