\name{density.gradient}
\alias{density.gradient}
\title{density gradient}
\description{Calculate a 2d density estimate then follow the gradient to a
point outside the convex hull.}
\usage{
density.gradient(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





