\name{direct.label}
\Rdversion{1.1}
\alias{direct.label}
\title{direct label}
\description{Add direct labels to a plot. This is a S3 generic and there are appropriate methods for "trellis" and "ggplot" objects.}
\usage{
direct.label(p, method = NULL, debug = FALSE)
}
\arguments{
  \item{p}{The plot to which you would like to add direct labels.}
  \item{method}{The direct label placement method as described in ?label.positions.}
  \item{debug}{Show debug output?}
}
\details{
}
\value{The plot object, with direct labels added.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}