\name{directlabels-package}
\alias{directlabels-package}
\docType{package}
\title{Direct labels for plots}
\description{Direct labeling functions meant to be used with several plotting packages, including lattice and ggplot2.}
\details{
\tabular{ll}{Package: \tab directlabels\cr}
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}

\keyword{ package }


