\name{dl.summarize}
\alias{dl.summarize}
\title{dl summarize}
\description{summarize which preserves important columns for direct labels.}
\usage{dl.summarize(OLD, ...)}
\arguments{
  \item{OLD}{data frame}
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





