\name{dldoc}
\alias{dldoc}
\title{Make directlabels documentation}
\description{Positioning Methods for direct labels are supposed to work with
only certain plot types. Each Positioning Method is defined in
R/file.R and plot examples are found in tests/doc/file/*.R so that
we can automatically assemble a database of example plots from the
code.}
\usage{dldoc(pkgdir = "..")}
\arguments{
  \item{pkgdir}{Package directory root.}
}

\value{Matrix of lists describing example plots and matching builtin
Positioning Methods.}

\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





