\name{dlgrob}
\alias{dlgrob}
\title{dlgrob}
\description{Make a grid grob that will draw direct labels.}
\usage{dlgrob(data, method, 
    debug = FALSE, axes2native = identity, 
    ...)}
\arguments{
  \item{data}{Data frame including points to plot in native coordinates.}
  \item{method}{Positioning Method.}
  \item{debug}{debug }
  \item{axes2native}{axes2native }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





