\name{draw.polygons}
\alias{draw.polygons}
\title{draw polygons}
\description{Draw polygons around label positions.}
\usage{draw.polygons(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





