\name{draw.rects}
\alias{draw.rects}
\title{draw rects}
\description{Positioning Function that draws boxes around label positions. Need
to have previously called \code{\link{calc.boxes}}. Does not edit the data
frame.}
\usage{draw.rects(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





