\name{edges.to.outside}
\alias{edges.to.outside}
\title{edges to outside}
\description{Given a list of \code{edges} from the convex or alpha hull, and a list of
cluster \code{centers}, calculate a point near to each cluster on the
outside of the hull.}
\usage{edges.to.outside(edges, 
    centers, debug = FALSE, 
    ...)}
\arguments{
  \item{edges}{
}
  \item{centers}{
}
  \item{debug}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





