\name{empty.grid.2}
\alias{empty.grid.2}
\title{empty grid 2}
\description{Use the perpendicular lines method in combination with the empty
grid method.}
\usage{
empty.grid.2(d, debug, ...)
}
\arguments{
  \item{d}{Data frame with columns groups x y.}
  \item{debug}{Show debugging graphics on the plot?}
  \item{\dots}{ignored.}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}