\name{empty.grid.collide}
\alias{empty.grid.collide}
\title{empty grid collide}
\description{Start with points from empty.grid then attempt to adjust labels so
that they do not collide with any points.}
\usage{
empty.grid.collide(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}
