\name{enlarge.box}
\alias{enlarge.box}
\title{enlarge box}
\description{Make text bounding box larger by some amount.}
\usage{enlarge.box(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





