\name{extreme.points}
\alias{extreme.points}
\title{extreme points}
\description{Label the points furthest from the middle for each group.}
\usage{extreme.points(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





