\name{filltemplate}
\alias{filltemplate}
\title{filltemplate}
\description{Fill in occurances of OBJ$item in the file \code{template} with the value
in R of L$item.}
\usage{filltemplate(L, template)}
\arguments{
  \item{L}{L }
  \item{template}{template }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





