\name{first.bumpup}
\alias{first.bumpup}
\docType{data}
\title{first bumpup}
\description{Label first points, bumping labels up if they collide.}
\usage{first.bumpup}
\format{
  The format is:
List of 2
 $ :function (d, ...)  
  ..- attr(*, "source")= chr "dl.indep(data.frame(d[which.min(d$x), ], hjust = 1, vjust = 0.5))"
 $ :function (d, ...)  
  ..- attr(*, "source")= chr [1:13] "function(d,...)" ...
}




