\name{first.points}
\alias{first.points}
\title{first points}
\description{Positioning Method for the first of a group of points.}
\usage{first.points(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





