\name{first.qp}
\alias{first.qp}
\docType{data}
\title{first qp}
\description{Label first points from QP solver that ensures labels do not collide.}
\usage{first.qp}
\format{
  The format is:
List of 3
 $ :function (d, ...)  
  ..- attr(*, "source")= chr "dl.indep(data.frame(d[which.min(d$x), ], hjust = 1, vjust = 0.5))"
 $ :function (d, debug = FALSE, ...)  
  ..- attr(*, "source")= chr [1:18] "function(d,debug=FALSE,...)" ...
 $ :function (d, ...)  
  ..- attr(*, "source")= chr [1:14] "function(d,...)" ...
}




