\name{first.smart}
\alias{first.smart}
\docType{data}
\title{first smart}
\description{Label first points but make sure labels do not collide.}
\usage{first.smart}
\format{
  The format is:
List of 2
 $ :function (d, ...)  
  ..- attr(*, "source")= chr "dl.indep(data.frame(d[which.min(d$x), ], hjust = 1, vjust = 0.5))"
 $ :function (d, ...)  
  ..- attr(*, "source")= chr [1:13] "function(d,...)" ...
}
\details{
}
\source{
}
\references{
}
\examples{}
