\name{follow.points}
\alias{follow.points}
\title{follow points}
\description{Draws a line between each center and every point, then follows the
line out far enough to give a box outside the cloud. Out of all
the boxes constructed in this way that do not contain any points,
take the one which has the smallest distance to the center. }
\usage{
follow.points(d,
debug = FALSE,
...)
}
\arguments{
  \item{d}{
}
  \item{debug}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





