\name{get.means}
\alias{get.means}
\title{get means}
\description{Positioning Function for the mean of each cluster of points.}
\usage{get.means(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





