\name{getLegendVariables}
\alias{getLegendVariables}
\title{getLegendVariables}
\description{get the aes which are variable in one legend.}
\usage{getLegendVariables(mb)}
\arguments{
  \item{mb}{mb }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





