\name{group.chull}
\alias{group.chull}
\title{group chull}
\description{Put the labels just outside the closest point in the union of the
group convex hulls. NEED TO find the point closest to the exterior
of the data.}
\usage{
group.chull(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





