\name{ignore.na}
\alias{ignore.na}
\title{ignore na}
\description{Remove rows for which either x or y is NA}
\usage{ignore.na(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





