\name{in1box}
\alias{in1box}
\title{in1box}
\description{Calculate how many points fall in a box.}
\usage{
in1box(p, box)
}
\arguments{
  \item{p}{data frame of points with columns x and y and many rows.}
  \item{box}{data frame of 1 row with columns left right top bottom.}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}
