\name{in1which}
\alias{in1which}
\title{in1which}
\description{Calculate which points fall in a \code{box}.}
\usage{in1which(p, box)}
\arguments{
  \item{p}{data frame of points with columns x and y and many rows.}
  \item{box}{data frame of 1 row with columns left right top bottom.}
}



\author{Toby Dylan Hocking}





