\name{label.pieces}
\alias{label.pieces}
\title{label pieces}
\description{Make a Positioning Method that will, for every piece, select
points and assign a vjust value.}
\usage{label.pieces(FUN, VJUST)}
\arguments{
  \item{FUN}{FUN }
  \item{VJUST}{VJUST }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





