\name{lasso.labels}
\alias{lasso.labels}
\docType{data}
\title{lasso labels}
\description{Label points at the zero before the first nonzero y value.}
\usage{lasso.labels}
\format{
  The format is:
List of 7
 $ rot: num 60
 $    :function (d, ...)  
  ..- attr(*, "source")= chr [1:6] "dl.indep()" ...
 $    :function (d, debug = FALSE, ...)  
 $    :function (d, ...)  
 $    :function (d, ...)  
 $    :function (d, ...)  
 $    :function (d, ...)  
}




