\name{last.bumpup}
\alias{last.bumpup}
\docType{data}
\title{last bumpup}
\description{Label last points, bumping labels up if they collide.}
\usage{last.bumpup}
\format{
  The format is:
List of 2
 $ :function (d, ...)  
  ..- attr(*, "source")= chr "dl.indep(data.frame(d[which.max(d$x), ], hjust = 0, vjust = 0.5))"
 $ :function (d, ...)  
}




