\name{last.qp}
\alias{last.qp}
\docType{data}
\title{last qp}
\description{Label last points from QP solver that ensures labels do not collide.}
\usage{last.qp}
\format{
  The format is:
List of 2
 $ :function (d, ...)  
  ..- attr(*, "source")= chr "dl.indep(data.frame(d[which.max(d$x), ], hjust = 0, vjust = 0.5))"
 $ :List of 2
  ..$ :function (d, ...)  
  ..$ :function (d, ...)  
}
\details{
}
\source{
}
\references{
}
\examples{}
