\name{last.smart}
\alias{last.smart}
\docType{data}
\title{last smart}
\description{Label last points but make sure labels do not collide.}
\usage{last.smart}
\format{
  The format is:
List of 2
 $ :function (d, ...)  
  ..- attr(*, "source")= chr "dl.indep(data.frame(d[which.max(d$x), ], hjust = 0, vjust = 0.5))"
 $ :function (d, ...)  
  ..- attr(*, "source")= chr [1:13] "function(d,...)" ...
}
\details{
}
\source{
}
\references{
}
\examples{}
