\name{lattice.translators}
\alias{lattice.translators}
\docType{data}
\title{lattice translators}
\description{Some lattice plot functions do some magic in the background to
translate the data you give them into the data points that are
plotted onscreen. We have to replicate this magic in native
coordinate space before applying the Positioning Method in cm
space. These functions accomplish this translation.}
\usage{"lattice.translators"}





