\name{left.points}
\alias{left.points}
\title{left points}
\description{Positioning Method for the first of a group of points.}
\usage{left.points(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





