\name{legends2hide}
\alias{legends2hide}
\title{legends2hide}
\description{Extract guides to hide from a ggplot.}
\usage{legends2hide(p)}
\arguments{
  \item{p}{p }
}

\value{NULL if no legends with colour or fill to hide.}

\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





