\name{make.tiebreaker}
\alias{make.tiebreaker}
\title{make tiebreaker}
\description{Make a tiebreaker function that can be used with \code{\link{qp.labels}}.}
\usage{make.tiebreaker(x.var, 
    tiebreak.var)}
\arguments{
  \item{x.var}{x.var }
  \item{tiebreak.var}{tiebreak.var }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





