\name{maxvar.points}
\alias{maxvar.points}
\title{maxvar points}
\description{Do first or last, whichever has points most spread out.}
\usage{maxvar.points(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





