\name{merge_recurse}
\alias{merge_recurse}
\title{merge recurse}
\description{Copied from reshape.}
\usage{merge_recurse(dfs, ...)}
\arguments{
  \item{dfs}{dfs }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





