\name{odd_timings}
\alias{odd_timings}
\docType{data}
\title{
  Odd timings
}
\description{
  These timings data made strange output labels with the
  "right.polygons" method.
}
\usage{data("odd_timings")}
\format{
  A data frame with 116 observations on the following 4 variables. Plot
  median.seconds versus N.col using a different line for each fun and a
  different panel for each captures.
  \describe{
    \item{\code{N.col}}{a numeric vector}
    \item{\code{fun}}{a character vector}
    \item{\code{captures}}{a numeric vector}
    \item{\code{median.seconds}}{a numeric vector}
  }
}
\source{
  \url{https://github.com/tdhock/nc-article}
}
\keyword{datasets}
