\name{only.unique.vals}
\alias{only.unique.vals}
\title{only unique vals}
\description{Create a 1-row data.frame consisting of only the columns for which
there is only 1 unique value.}
\usage{only.unique.vals(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





