\name{outside.ahull}
\alias{outside.ahull}
\title{outside ahull}
\description{Calculate closest point on the alpha hull with size of the boxes,
and put it outside that point.}
\usage{outside.ahull(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





