\name{outside.chull}
\alias{outside.chull}
\title{outside chull}
\description{Calculate closest point on the convex hull and put it outside that
point. Assume \code{d} is the center for each point cloud and then use
orig.data to calculate hull.}
\usage{outside.chull(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





