\name{pkgFun}
\alias{pkgFun}
\title{pkgFun}
\description{https://github.com/tdhock/directlabels/issues/2 CRAN won't
complain about this version of :::}
\usage{pkgFun(fun, pkg = "ggplot2")}
\arguments{
  \item{fun}{fun }
  \item{pkg}{pkg }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





