\name{position_dl}
\alias{position_dl}
\title{position dl}
\description{Position for internal use with geom_text.}
\usage{
position_dl(method = NULL, debug = FALSE, orig = NULL, ...)
}
\arguments{
  \item{method}{
}
  \item{debug}{
}
  \item{orig}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





