\name{PositionDl}
\alias{PositionDl}
\docType{data}
\title{PositionDl}
\description{Position class for direct label placement.}
\usage{PositionDl}
\format{
  The format is:
proto object 
 $ objname: chr "dl" 
 $ adjust :function (., data, scales)   
  ..- attr(*, "source")= chr [1:32] "function(.,data,scales)" ... 
 $ new    :function (., method = NULL, debug = FALSE, orig = NULL)   
  ..- attr(*, "source")= chr [1:3] "function(., method=NULL, debug=FALSE, orig=NULL) " ... 
 $ orig   : NULL 
 $ debug  : logi FALSE 
 $ method : NULL 
 parent: proto object 
 .. parent: proto object 
}




