\name{reduce.cex.tb}
\alias{reduce.cex.tb}
\title{reduce cex tb}
\description{If edges of the text are going over the top or bottom of the
plotting region, then decrease cex until it fits.}
\usage{reduce.cex.tb(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





