\name{right.points}
\alias{right.points}
\title{right points}
\description{Positioning Method for the last of a group of points.}
\usage{right.points(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





