\name{rug.mean}
\alias{rug.mean}
\title{rug mean}
\description{Place points on top of the mean value of the rug.}
\usage{
rug.mean(d, ..., end)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
  \item{end}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}