\name{static.labels}
\alias{static.labels}
\title{static labels}
\description{to hard-code label positions...}
\usage{static.labels(x, y, groups, 
    ...)}
\arguments{
  \item{x}{x }
  \item{y}{y }
  \item{groups}{groups }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





