\name{top.bumptwice}
\alias{top.bumptwice}
\title{top bumptwice}
\description{Label the tops, bump labels up to avoid other labels, then to the
side to avoid collisions with points.}
\usage{top.bumptwice(d, debug = FALSE, 
    ...)}
\arguments{
  \item{d}{d }
  \item{debug}{debug }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





