\name{top.points}
\alias{top.points}
\title{top points}
\description{Positioning Method for the top of a group of points.}
\usage{top.points(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





