\name{top.qp}
\alias{top.qp}
\docType{data}
\title{top qp}
\description{Label points at the top, making sure they don't collide.}
\usage{top.qp}
\format{
  The format is:
List of 2
 $ :function (d, ...)  
  ..- attr(*, "source")= chr "dl.indep(data.frame(d[which.max(d$y), ], hjust = 0.5, vjust = 0))"
 $ :List of 2
  ..$ :function (d, ...)  
  .. ..- attr(*, "source")= chr [1:21] "function(d,...)" ...
  ..$ :function (d, ...)  
  .. ..- attr(*, "source")= chr [1:13] "function(d,...)" ...
}
\details{
}
\source{
}
\references{
}
\examples{}
