\name{trans.density}
\Rdversion{1.1}
\alias{trans.density}
\title{trans density}
\description{Transformation function for 1d densityplots.}
\usage{
trans.density(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}