\name{trans.densityplot}
\alias{trans.densityplot}
\title{trans densityplot}
\description{Transformation function for 1d densityplots.}
\usage{
trans.densityplot(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





