\name{trans.qqmath}
\alias{trans.qqmath}
\title{trans qqmath}
\description{Transformation function for 1d qqmath plots. This is a copy-paste
from panel.qqmath. (total hack)}
\usage{
trans.qqmath(d, distribution, f.value, qtype = 7, ...)
}
\arguments{
  \item{d}{
}
  \item{distribution}{
}
  \item{f.value}{
}
  \item{qtype}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





