\name{uselegend.ggplot}
\alias{uselegend.ggplot}
\title{uselegend ggplot}
\description{Show the ggplot2 legend, for comparison.}
\usage{uselegend.ggplot(p, ...)}
\arguments{
  \item{p}{The ggplot object.}
  \item{\dots}{Ignored.}
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





