\name{uselegend.trellis}
\alias{uselegend.trellis}
\title{uselegend trellis}
\description{Add a legend to a trellis plot, for comparison.}
\usage{uselegend.trellis(p, 
    ...)}
\arguments{
  \item{p}{The trellis object.}
  \item{\dots}{Ignored.}
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





