\name{vertical.qp}
\alias{vertical.qp}
\title{vertical qp}
\description{Make a Positioning Function from a set of points on a vertical
line that will be spaced out using \code{\link{qp.labels}}.}
\usage{vertical.qp(M)}
\arguments{
  \item{M}{M }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





