\name{visualcenter}
\alias{visualcenter}
\title{visualcenter}
\description{Point in the middle of the min and max for each group.}
\usage{visualcenter(d, ...)}
\arguments{
  \item{d}{d }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





