\name{xlimits}
\alias{xlimits}
\title{xlimits}
\description{Return the positions of the plot horizontal limits in cm, for use
as the limit argument to \code{\link{qp.labels}}.}
\usage{xlimits(...)}
\arguments{
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





