\name{ylimits}
\alias{ylimits}
\title{ylimits}
\description{Return the positions of the plot vertical limits in cm, for use as
the limit argument to \code{\link{qp.labels}}.}
\usage{ylimits(...)}
\arguments{
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





