\name{bottom.points}
\Rdversion{1.1}
\alias{bottom.points}
\title{bottom points}
\description{Positioning Function for the bottom of a group of points.}
\usage{
bottom.points(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}