\name{bumpup}
\alias{bumpup}
\title{bumpup}
\description{Sequentially bump labels up, starting from the bottom, if they
collide with the label underneath. NOTE: behavior is undefined
when used with ggplot2 since it relies on the calc.boxes()
function which doesn't know how to calculate bounding boxes for
ggplot2 labels (yet).}
\usage{
bumpup(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





