\name{calc.boxes}
\alias{calc.boxes}
\title{calc boxes}
\description{Calculate boxes around labels, for collision detection.}
\usage{
calc.boxes(d, debug = FALSE, class = "ggplot2", ...)
}
\arguments{
  \item{d}{
}
  \item{debug}{
}
  \item{class}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





