\name{closest.on.ahull}
\alias{closest.on.ahull}
\title{closest on ahull}
\description{Calculate closest point on the alpha hull with size of the boxes,
and put it outside that point. (only technically correct for
aspect="iso" TODO: check and correct for perspective changes.)
TODO: doesn't work with ggplot2 since we can't calculate bounding
box.}
\usage{
closest.on.ahull(d, debug = FALSE, center.fun = big.boxes, ...)
}
\arguments{
  \item{d}{
}
  \item{debug}{
}
  \item{center.fun}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





