\name{closest.on.chull}
\alias{closest.on.chull}
\title{closest on chull}
\description{Calculate closest point on the convex hull and put it outside that
point.}
\usage{
closest.on.chull(d, debug = FALSE, center.fun = big.boxes, ...)
}
\arguments{
  \item{d}{
}
  \item{debug}{
}
  \item{center.fun}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





