\name{default.ahull}
\alias{default.ahull}
\title{default ahull}
\description{Calculate the default alpha parameter for ashape based on the
average size of label boxes.}
\usage{default.ahull(d, ...)}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





