\name{defaultpf.ggplot}
\alias{defaultpf.ggplot}
\title{defaultpf ggplot}
\description{Default method selection method for ggplot2 plots.}
\usage{
defaultpf.ggplot(geom,
p,
...)
}
\arguments{
  \item{geom}{
}
  \item{p}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





