\name{defaultpf.ggplot}
\alias{defaultpf.ggplot}
\title{defaultpf ggplot}
\description{Default method selection method for ggplot2 plots.}
\usage{defaultpf.ggplot(geom, 
    p, L, colvar, ...)}
\arguments{
  \item{geom}{geom }
  \item{p}{p }
  \item{L}{L }
  \item{colvar}{colvar }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre] (ORCID: <https://orcid.org/0000-0002-3146-0865>)}





