\name{defaultpf.ggplot}
\alias{defaultpf.ggplot}
\title{defaultpf ggplot}
\description{Default method selection method for ggplot2 plots.}
\usage{defaultpf.ggplot(geom, p, L, colvar, ...)}
\arguments{
  \item{geom}{
}
  \item{p}{
}
  \item{L}{
}
  \item{colvar}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





