\name{defaultpf.trellis}
\alias{defaultpf.trellis}
\title{defaultpf trellis}
\description{If no Positioning Method specified, choose a default using this
function. The idea is that this is called with all the variables
in the environment of panel.superpose.dl, and this can be
user-customizable by setting the directlabels.defaultpf.lattice
option to a function like this.}
\usage{defaultpf.trellis(lattice.fun.name, groups, type, ...)}
\arguments{
  \item{lattice.fun.name}{
}
  \item{groups}{
}
  \item{type}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





