\name{direct.label.ggplot}
\alias{direct.label.ggplot}
\title{direct label ggplot}
\description{Direct label a ggplot2 grouped plot.}
\usage{
direct.label.ggplot(p, method = NULL, debug = FALSE)
}
\arguments{
  \item{p}{The ggplot object.}
  \item{method}{Method for direct labeling as described in ?label.positions.}
  \item{debug}{Show debug output?}
}
\details{
}
\value{The ggplot object with direct labels added.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}
