\name{direct.label.ggplot}
\alias{direct.label.ggplot}
\title{direct label ggplot}
\description{Direct label a ggplot2 grouped plot.}
\usage{\method{direct.label}{ggplot}(p, 
    method = NULL, debug = FALSE)}
\arguments{
  \item{p}{The ggplot object.}
  \item{method}{Method for direct labeling as described in
\code{\link{apply.method}}.}
  \item{debug}{Show \code{debug} output?}
}

\value{The ggplot object with direct labels added.}

\author{Toby Dylan Hocking}





