\name{directlabels-package}
\alias{directlabels-package}
\docType{package}
\title{Direct labels for multicolor plots in lattice or ggplot2}
\description{The directlabels package provides an extensible framework
for automatically placing direct labels onto multicolor lattice or
ggplot2 plots.
It includes heuristics for examining "lattice" and "ggplot" objects
and inferring an appropriate Positioning Method for placing the
labels. Furthermore, the design of directlabels makes it simple to
create Positioning Methods for specific plots or libraries of
portable Positioning Methods that can be re-used.}
\details{
\tabular{ll}{Package: \tab directlabels\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@inria.fr>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 2.0\cr
License: \tab GPL-3\cr
Title: \tab Direct labels for multicolor plots in lattice or ggplot2\cr
URL: \tab http://directlabels.r-forge.r-project.org/\cr
LazyData: \tab true\cr
Suggests: \tab ks, MASS, inlinedocs, quadprog, ggplot2, lattice, alphahull,
ElemStatLearn, proto, lars, mlmRev, latticeExtra\cr
Depends: \tab grid\cr
Collate: \tab utility.function.R compare.R dotplot.R lineplot.R
densityplot.R  ggplot2.R  positioning.functions.R
doc.R lattice.R scatterplot.R contourplot.R\cr}
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}

\keyword{ package }


