\name{directlabels-package}
\Rdversion{1.1}
\alias{directlabels-package}
\docType{package}
\title{Direct labels for plots}
\description{Direct labeling functions meant to be used with several plotting packages, including lattice and ggplot2.}
\details{
\tabular{ll}{Package: \tab directlabels\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@inria.fr>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 1.0\cr
License: \tab GPL-3\cr
Title: \tab Direct labels for plots\cr
Description: \tab Direct labeling functions meant to be used with several plotting packages, including lattice and ggplot2.\cr
Imports: \tab proto, plyr, lattice, grid, ggplot2\cr
Depends: \tab proto, plyr, lattice, grid, ggplot2\cr
URL: \tab http://directlabels.r-forge.r-project.org/\cr
LazyLoad: \tab false\cr
LazyData: \tab true\cr
Suggests: \tab nlme, mlmRev\cr}
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\references{
}
\keyword{ package }
\seealso{
}
\examples{}