\name{directlabels-package}
\alias{directlabels-package}
\docType{package}
\title{Direct labels for plots}
\description{Direct labeling functions meant to be used with several plotting packages, including lattice and ggplot2.}
\details{
\tabular{ll}{Package: \tab directlabels\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@inria.fr>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 1.0\cr
License: \tab GPL-3\cr
Title: \tab Direct labels for plots\cr
Imports: \tab proto, plyr, lattice, grid, ggplot2 (>= 0.8.5)\cr
Depends: \tab proto, plyr, lattice, grid, ggplot2 (>= 0.8.5)\cr
URL: \tab http://directlabels.r-forge.r-project.org/\cr
LazyLoad: \tab false\cr
LazyData: \tab true\cr
Suggests: \tab nlme, mlmRev, ElemStatLearn, inlinedocs, quadprog, lars, ks\cr
Collate: \tab utility.function.R compare.R dotplot.R lineplot.R
densityplot.R  ggplot2.R  positioning.functions.R
doc.R lattice.R scatterplot.R contourplot.R\cr}
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}

\keyword{ package }


