\name{directlabels-package}
\alias{directlabels-package}
\alias{directlabels}
\docType{package}
\title{Direct labels for multicolor plots in lattice or ggplot2}
\description{An extensible framework
for automatically placing direct labels onto multicolor lattice or
ggplot2 plots.
Label positions are described using Positioning Methods
which can be re-used across several different plots.
There are heuristics for examining "trellis" and "ggplot" objects
and inferring an appropriate Positioning Method.}
\details{
\tabular{ll}{Package: \tab directlabels\cr
Maintainer: \tab Toby Dylan Hocking <toby@sg.cs.titech.ac.jp>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 2014.6.13\cr
License: \tab GPL-3\cr
Title: \tab Direct labels for multicolor plots in lattice or ggplot2\cr
URL: \tab http://directlabels.r-forge.r-project.org/\cr
LazyData: \tab true\cr
Suggests: \tab MASS, inlinedocs, ggplot2 (>= 0.9.1),
lattice, alphahull, reshape2, nlme,
ElemStatLearn, proto, lars, latticeExtra\cr
Depends: \tab grid, quadprog\cr
Collate: \tab utility.function.R compare.R dotplot.R lineplot.R
densityplot.R  ggplot2.R  positioning.functions.R
doc.R lattice.R scatterplot.R contourplot.R\cr}
}
\author{Toby Dylan Hocking}

\keyword{ package }


