\name{dl.summarize}
\alias{dl.summarize}
\title{dl summarize}
\description{summarize which preserves important columns for direct labels.}
\usage{
dl.summarize(d,
...)
}
\arguments{
  \item{d}{data frame}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





