\name{dlgrob}
\alias{dlgrob}
\title{dlgrob}
\description{Make a grid grob that will draw direct labels.}
\usage{
dlgrob(data, method, debug = FALSE, ...)
}
\arguments{
  \item{data}{Data frame including points to plot in native coordinates.}
  \item{method}{Positioning Method.}
  \item{debug}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





