\name{dlgrob}
\alias{dlgrob}
\title{dlgrob}
\description{Make a grid grob that will draw direct labels.}
\usage{dlgrob(data, method, debug = FALSE, axes2native = identity, ...)}
\arguments{
  \item{data}{Data frame including points to plot in native coordinates.}
  \item{method}{Positioning Method.}
  \item{debug}{
}
  \item{axes2native}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





