\name{draw.rects}
\alias{draw.rects}
\title{draw rects}
\description{Positioning Function that draws boxes around label positions. Need
to have previously called \code{\link{calc.boxes}}. Does not edit the data
frame.}
\usage{draw.rects(d, ...)}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





