\name{drawDetails.dlgrob}
\alias{drawDetails.dlgrob}
\title{drawDetails dlgrob}
\description{Process data points using the Positioning Method and draw the
resulting direct labels. This is called for every panel with
direct labels, every time the plot window is resized.}
\usage{
\method{drawDetails}{dlgrob}(x,
recording)
}
\arguments{
  \item{x}{
}
  \item{recording}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





