\name{enlarge.box}
\alias{enlarge.box}
\title{enlarge box}
\description{Make text bounding box larger by some amount.}
\usage{
enlarge.box(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}
