\name{extract.plot}
\alias{extract.plot}
\title{Extract plot and definition for documentation}
\description{Given an R code file, execute it, store the definition, and save
the resulting plot in a variable.}
\usage{extract.plot(f)}
\arguments{
  \item{f}{R code file with plot example.}
}



\author{Toby Dylan Hocking}





