\name{far.from.others.borders}
\alias{far.from.others.borders}
\title{far from others borders}
\description{Find the point on each curve which maximizes the distance to the
plot border or to another curve.}
\usage{far.from.others.borders(all.groups, 
    ..., debug = FALSE)}
\arguments{
  \item{all.groups}{
}
  \item{\dots}{
}
  \item{debug}{
}
}



\author{Toby Dylan Hocking}





