\name{first.points}
\alias{first.points}
\title{first points}
\description{Positioning Function for the first of a group of points.}
\usage{
first.points(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}
