\name{first.qp}
\alias{first.qp}
\docType{data}
\title{first qp}
\description{Label first points from QP solver that ensures labels do not collide.}
\usage{first.qp}
\format{
  The format is:
List of 2
 $ :function (d, ...)  
  ..- attr(*, "source")= chr "dl.indep(data.frame(d[which.min(d$x), ], hjust = 1, vjust = 0.5))"
 $ :List of 2
  ..$ :function (d, ...)  
  ..$ :function (d, ...)  
}




