\name{lasso.labels}
\alias{lasso.labels}
\docType{data}
\title{lasso labels}
\description{Label points at the zero before the first nonzero y value.}
\usage{lasso.labels}
\format{
  The format is:
List of 7
 $ rot: num 60
 $    :function (d, ...)  
  ..- attr(*, "source")= chr "dl.indep(\n    d <- d[order(d$x), ]\n    i <- which(d$y != 0)[1]\n    hjust <- as.integer(d[i, \"y\"] > 0)\n    data.frame(d[i"| __truncated__
 $    :function (d, debug = FALSE, ...)  
  ..- attr(*, "source")= chr [1:18] "function(d,debug=FALSE,...)" ...
 $    :function (d, ...)  
  ..- attr(*, "source")= chr "function(d,...)do.call(\"transform\",c(list(d),L))"
 $    :function (d, ...)  
  ..- attr(*, "source")= chr "function(d,...)do.call(\"transform\",c(list(d),L))"
 $    :function (d, ...)  
  ..- attr(*, "source")= chr "function(d,...)do.call(\"transform\",c(list(d),L))"
 $    :function (d, ...)  
  ..- attr(*, "source")= chr [1:14] "function(d,...)" ...
}




