\name{make.tiebreaker}
\alias{make.tiebreaker}
\title{make tiebreaker}
\description{Make a tiebreaker function that can be used with \code{\link{qp.labels}}.}
\usage{make.tiebreaker(x.var, 
    tiebreak.var)}
\arguments{
  \item{x.var}{
}
  \item{tiebreak.var}{
}
}



\author{Toby Dylan Hocking}





