\name{maxvar.points}
\alias{maxvar.points}
\title{maxvar points}
\description{Do first or last, whichever has points most spread out.}
\usage{
maxvar.points(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}