\name{only.unique.vals}
\alias{only.unique.vals}
\title{only unique vals}
\description{Create a 1-row data.frame consisting of only the columns for which
there is only 1 unique value.}
\usage{only.unique.vals(d, ...)}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





