\name{outside.ahull}
\alias{outside.ahull}
\title{outside ahull}
\description{Calculate closest point on the alpha hull with size of the boxes,
and put it outside that point.}
\usage{outside.ahull(d, ...)}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





