\name{outside.chull}
\alias{outside.chull}
\title{outside chull}
\description{Calculate closest point on the convex hull and put it outside that
point. Assume \code{d} is the center for each point cloud and then use
orig.data to calculate hull.}
\usage{outside.chull(d, ...)}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





