\name{polygon.method}
\alias{polygon.method}
\title{polygon method}
\description{Make a Positioning Method that places non-overlapping speech
polygons at the first or last points.}
\usage{polygon.method(top.bottom.left.right, 
    space = 0.1)}
\arguments{
  \item{top.bottom.left.right}{
}
  \item{space}{
}
}



\author{Toby Dylan Hocking}





