\name{polygon.method}
\alias{polygon.method}
\title{polygon method}
\description{Make a Positioning Method that places non-overlapping speech
polygons at the first or last points.}
\usage{polygon.method(method, space, data.col, na.col)}
\arguments{
  \item{method}{
}
  \item{space}{
}
  \item{data.col}{
}
  \item{na.col}{
}
}



\author{Toby Dylan Hocking}





