\name{position_dl}
\Rdversion{1.1}
\alias{position_dl}
\title{position dl}
\description{Position for internal use with geom_text.}
\usage{
position_dl(method = NULL, debug = FALSE, ...)
}
\arguments{
  \item{method}{
}
  \item{debug}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}