\name{PositionDl}
\alias{PositionDl}
\docType{data}
\title{PositionDl}
\description{Position class for direct label placement.}
\usage{PositionDl}
\format{
  The format is:
proto object 
 $ objname: chr "dl" 
 $ adjust :function (., data, scales)   
  ..- attr(*, "source")= chr [1:21] "function(.,data,scales)" ... 
 $ new    :function (., method = NULL, debug = FALSE)   
  ..- attr(*, "source")= chr [1:3] "function(., method=NULL, debug=FALSE) " ... 
 $ debug  : logi FALSE 
 $ method : NULL 
 parent: proto object 
 .. parent: proto object 
}
\details{
}
\source{
}
\references{
}
\examples{}
