\name{PositionDl}
\Rdversion{1.1}
\alias{PositionDl}
\docType{data}
\title{PositionDl}
\description{Position class for direct label placement.}
\usage{PositionDl}
\format{
  The format is:
proto object 
 $ objname: chr "dl" 
 $ adjust :function (., data, scales)   
 $ new    :function (., method = NULL, debug = FALSE)   
 $ debug  : logi FALSE 
 $ method : NULL 
 parent: proto object 
 .. parent: proto object 
}
\details{
}
\source{
}
\references{
}
\examples{}