\name{qp.labels}
\alias{qp.labels}
\title{qp labels}
\description{Use a QP solver to find the best places to put the points on a
line, subject to the constraint that they should not overlap
(assumes they all have the same height/width).}
\usage{
qp.labels(var, spacer)
}
\arguments{
  \item{var}{
}
  \item{spacer}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}
