\name{reduce.cex.lr}
\alias{reduce.cex.lr}
\title{reduce cex lr}
\description{If left or right edges of the text are going out of the plotting
region, then decrease cex until it fits.}
\usage{reduce.cex.lr(d, ...)}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





