\name{reduce.cex.tb}
\alias{reduce.cex.tb}
\title{reduce cex tb}
\description{If edges of the text are going over the top or bottom of the
plotting region, then decrease cex until it fits.}
\usage{reduce.cex.tb(d, ...)}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





