\name{rhtmlescape}
\alias{rhtmlescape}
\title{rhtmlescape}
\description{for standards compliance we should escape <>&}
\usage{
rhtmlescape(code)
}
\arguments{
  \item{code}{R code to be displayed on a HTML page between pre tags.}
}
\details{
}
\value{Standards compliant HTML to display.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}
