\name{rhtmlescape}
\alias{rhtmlescape}
\title{rhtmlescape}
\description{for standards compliance we should escape <>&}
\usage{rhtmlescape(code)}
\arguments{
  \item{code}{R \code{code} to be displayed on a HTML page between pre tags.}
}

\value{Standards compliant HTML to display.}

\author{Toby Dylan Hocking}





