\name{right.points}
\Rdversion{1.1}
\alias{right.points}
\title{right points}
\description{  dl.indep(data.frame(d[which.max(d$x),],hjust=0,vjust=0.5))}
\usage{
right.points(d, ...)
}
\arguments{
  \item{d}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}