\name{smart.grid}
\alias{smart.grid}
\docType{data}
\title{smart grid}
\description{Use bounding box information with a small empty.grid to find the a
non-colliding label that is close to a point on the convex hull,
which is close to the visual center of the data. TODO: does not
work with ggplot2 since the backend does not support bounding box
calculation.}
\usage{smart.grid}





