\name{smart.grid}
\alias{smart.grid}
\title{smart grid}
\description{Use bounding box information with a small empty.grid. TODO: does
not work with ggplot2 since the backend does not support bounding
box calculation.}
\usage{
smart.grid(d, debug, ...)
}
\arguments{
  \item{d}{
}
  \item{debug}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





