\name{top.bumptwice}
\alias{top.bumptwice}
\title{top bumptwice}
\description{Label the tops, bump labels up to avoid other labels, then to the
side to avoid collisions with points.}
\usage{top.bumptwice(d, debug = FALSE, ...)}
\arguments{
  \item{d}{
}
  \item{debug}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





