\name{top.qp}
\alias{top.qp}
\docType{data}
\title{top qp}
\description{Label points at the top, making sure they don't collide.}
\usage{top.qp}
\format{
  The format is:
List of 3
 $ :function (d, ...)  
  ..- attr(*, "source")= chr "dl.indep(data.frame(d[which.max(d$y), ], hjust = 0.5, vjust = 0))"
 $ :function (d, debug = FALSE, ...)  
 $ :function (d, ...)  
}




