\name{uselegend.trellis}
\alias{uselegend.trellis}
\title{uselegend trellis}
\description{Add a legend to a trellis plot, for comparison.}
\usage{
uselegend.trellis(p, ...)
}
\arguments{
  \item{p}{The trellis object.}
  \item{\dots}{Ignored.}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\note{
}


\seealso{
}
\examples{}
