\name{vertical.qp}
\alias{vertical.qp}
\title{vertical qp}
\description{Make a Positioning Function from a set of points on a vertical
line that will be spaced out using qp.labels}
\usage{vertical.qp(M, limits = NULL)}
\arguments{
  \item{M}{
}
  \item{limits}{
}
}



\author{Toby Dylan Hocking}





